+++ src/main/java/com/dodok/honeypot/domain/member/mapper/MemberMapper.java
@@ -2,7 +2,7 @@ package com.dodok.honeypot.domain.member.mapper;
 
 import com.dodok.honeypot.domain.member.dto.info.MemberInfo;
 import com.dodok.honeypot.domain.member.dto.res.MembersInfoResDto;
-import com.dodok.honeypot.global.entity.PageInfo;
+import com.dodok.honeypot.global.dto.PageInfo;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Component;
